Move AppScaffoldAction batck to scaffold.dart
